v1.20171014
I optimistically label this a stable release.

It's not actually complete or well-tested, but it's simple enough, and I
don't want to be too pedantic about it, 'cause I've got other things to
do, and this demonstrably works as much as we need it to.

Marking it as stable just means I hope I don't have to touch it again.